feat. adding option to enable ssh for hyperv tool

+++ tools/hyperv/create-dev-box.ps1
@@ -10,6 +10,9 @@
 
     .PARAMETER VMName
     VM Name used on VM creation, will also be used as the name of the .vhdx
+    
+    .PARAMETER EnableSSH
+    Enable the current user's ssh-pub-key as an authorized key on the VM
 
     .PARAMETER UserDataSource
     Path to source .yml for cloud-config user-data, see https://cloudinit.readthedocs.io/en/latest/reference/examples.html for examples
@@ -22,15 +25,17 @@ param (
     [Parameter(Mandatory)]
     [string]$VMName,
     [Parameter(Mandatory = $false)]
+    [bool]$EnableSSH,
+    [Parameter(Mandatory = $false)]
     [string]$UserDataSource = "./configs/overlaybd-dev.yml"
 )
 
 # Get the ID and security principal of the current user account
-$myWindowsID=[System.Security.Principal.WindowsIdentity]::GetCurrent()
-$myWindowsPrincipal=new-object System.Security.Principal.WindowsPrincipal($myWindowsID)
+$myWindowsID = [System.Security.Principal.WindowsIdentity]::GetCurrent()
+$myWindowsPrincipal = new-object System.Security.Principal.WindowsPrincipal($myWindowsID)
  
 # Get the security principal for the Administrator role
-$adminRole=[System.Security.Principal.WindowsBuiltInRole]::Administrator
+$adminRole = [System.Security.Principal.WindowsBuiltInRole]::Administrator
  
 # Check to see if we are currently running "as Administrator"
 if (!$myWindowsPrincipal.IsInRole($adminRole)) {
@@ -63,7 +68,8 @@ $lastModified = $manifestResponse.Headers.'Last-Modified'
 
 if ($PSVersionTable.PSVersion.Major -gt 5) {
     $lastModified = [DateTime]$lastModified[0]
-} else {
+}
+else {
     $lastModified = [DateTime]$lastModified
 }
 
@@ -74,7 +80,7 @@ instance-id: guid-$([GUID]::NewGuid())
 local-hostname: $($VMName)
 "@
 
-$networkconfig=@"
+$networkconfig = @"
 version: 2
 ethernets:
   eth0:
@@ -82,7 +88,7 @@ ethernets:
 "@
 
 # Check Paths
-if (!(test-path $imageCachePath)) {mkdir $imageCachePath}
+if (!(test-path $imageCachePath)) { mkdir $imageCachePath }
 
 if (!(test-path $vmConfigPath)) {
     mkdir -p $vmConfigPath
@@ -97,36 +103,68 @@ if (!(test-path $vmPath)) {
 }
 
 # Helper function for no error file cleanup
-Function cleanupFile ([string]$file) {if (test-path $file) {Remove-Item $file}}
+Function cleanupFile ([string]$file) { if (test-path $file) { Remove-Item $file } }
 
 # Create new virtual machine and start it
 # Delete the VM if it is around
 If ((Get-VM | Where-Object name -eq $VMName).Count -gt 0)
-      {Stop-VM $VMName -TurnOff -Confirm:$false -Passthru | Remove-VM -Force}
+{ Stop-VM $VMName -TurnOff -Confirm:$false -Passthru | Remove-VM -Force }
 cleanupFile $vhdx
 cleanupFile $metaDataIso
 
 # Make temp location
 if (!(test-path "$($imageCachePath)\ubuntu-$($stamp).img")) {
-      # If we do not have a matching image - delete the old ones and download the new one
-      Remove-Item "$($imageCachePath)\ubuntu-*.img"
-      Invoke-WebRequest "$($ubuntuPath).img" -UseBasicParsing -OutFile "$($imageCachePath)\ubuntu-$($stamp).img"
+    # If we do not have a matching image - delete the old ones and download the new one
+    Remove-Item "$($imageCachePath)\ubuntu-*.img"
+    Invoke-WebRequest "$($ubuntuPath).img" -UseBasicParsing -OutFile "$($imageCachePath)\ubuntu-$($stamp).img"
 }
 
 # Convert cloud image to VHDX
 & $qemuImgPath convert -f qcow2 "$($imageCachePath)\ubuntu-$($stamp).img" -O vhdx -o subformat=dynamic $vhdx
 
+if ($EnableSSH) {
+    $sshKey = Get-Content "c:\Users\$($env:USERNAME)\.ssh\id_rsa.pub"
+    $sshKey = $sshKey.Replace("\", "\\");
+
+    $vendorData = @"
+## template: jinja
+#cloud-config
+merge_how:
+ - name: list
+   settings: [append]
+ - name: dict
+   settings: [no_replace, recurse_list]
+
+write_files:
+ - content: | 
+      $($sshKey)
+   path: /home/chief/.ssh/authorized_keys
+"@
+
+    if ($PSVersionTable.PSVersion.Major -gt 5) {
+        # Output meta, network, and user data to files
+        Set-Content "$($nocloudPath)\vendor-data" ([byte[]][char[]] "$vendorData") -AsByteStream
+    }
+    else {
+        # Output meta, network, and user data to files
+        Set-Content "$($nocloudPath)\vendor-data" ([byte[]][char[]] "$vendorData") -Encoding Byte
+    }
+}
+
 if ($PSVersionTable.PSVersion.Major -gt 5) {
     # Output meta, network, and user data to files
     Set-Content "$($nocloudPath)\meta-data" ([byte[]][char[]] "$metadata") -AsByteStream
     Set-Content "$($nocloudPath)\network-config" ([byte[]][char[]] "$networkconfig") -AsByteStream
-} else {
+}
+else {
     # Output meta, network, and user data to files
     Set-Content "$($nocloudPath)\meta-data" ([byte[]][char[]] "$metadata") -Encoding Byte
     Set-Content "$($nocloudPath)\network-config" ([byte[]][char[]] "$networkconfig") -Encoding Byte
 }
 
-Set-Content "$($nocloudPath)\user-data" (Get-Content $UserDataSource)
+if (!(test-path "$($nocloudPath)\user-data")) {
+    Set-Content "$($nocloudPath)\user-data" (Get-Content $UserDataSource)
+}
 
 # Create meta data ISO image
 & $oscdimgPath $nocloudPath $metaDataIso -j2 -lcidata
@@ -134,7 +172,7 @@ Set-Content "$($nocloudPath)\user-data" (Get-Content $UserDataSource)
 Resize-VHD -Path $vhdx -SizeBytes 512GB
 
 New-VM $VMName -MemoryStartupBytes 4096mb -BootDevice VHD -VHDPath $vhdx -Generation 2 `
-               -SwitchName $virtualSwitchName -Path $vmPath | Out-Null
+    -SwitchName $virtualSwitchName -Path $vmPath | Out-Null
 Set-VM -Name $VMName -ProcessorCount 2 -AutomaticStopAction ShutDown -AutomaticStartAction StartIfRunning -AutomaticStartDelay (Get-Random -Minimum 100 -Maximum 800)
 Set-VMFirmware -VMName $VMName -EnableSecureBoot Off -FirstBootDevice (Get-VMHardDiskDrive -VMName $VMName)
 Get-VM -VMname $VMName | Enable-VMIntegrationService -Name *
